refactor: Migrated from Newtonsoft.Json to System.Text.Json, missing one test passing

+++ src/ElectronNET.API/API/Entities/CPUUsage.cs
@@ -8,7 +8,7 @@
         /// <summary>
         /// Percentage of CPU used since the last call to getCPUUsage. First call returns 0.
         /// </summary>
-        public int PercentCPUUsage { get; set; }
+        public double PercentCPUUsage { get; set; }
 
         /// <summary>
         /// The number of average idle cpu wakeups per second since the last call to
